CR5 - Bus gets dijkstra and works
Had a bug,
Bug was that Graph wasn't created correctly. It had duplicate nodes that did not connect correctly. Had to make the graph creation smarter, with more error checking.

Now I can create a slotpath for a bus, according to dijkstra will get the fastest route from A to B.

Also draws all the slotpositions.
Which have to be more later on. For now they are just one per road.